Add files via upload
This is a generalyzed script to compare multiple quantitative variables between groups (a.k.a. based on a qualitative or factorial variable). It performes serial histograms and boxplot across groups, finds outliers and displays a result table containing the means, sd and statistical significance for every group/factor of the different variables.

Includes a dummy dataset to try it.